[3 - 4단계 방탈출 예약 대기] 몰리(김지민) 미션 제출합니다. (#145)
* style: 요구사항에 맞게 프론트 코드 수정

* docs(README.md): 3단계 요구사항 작성

* docs(erd.png): 예약 대기 스키마 설계

* refactor(BaseEntity): 모든 엔티티들의 생성 시간, 수정 시간을 저장하는 BaseEntity 생성 및 상속

* feat(Waiting): 예약 대기에 대한 엔티티 구현

* docs(APISpecification): 예약 대기에 대한 API 명세 작성

* test(WaitingRepositoryTest): 해당하는 회원, 예약과 동일한 예약 대기가 존재하는지에 대한 쿼리 테스트 작성

* feat(WaitingController): 예약 대기 생성 기능 구현

* feat(Waiting, Member): 회원에 대한 예약인지 묻는 메서드 추가

* refactor(/exception): 401, 403 커스텀 예외 생성

* feat(WaitingController): 예약 대기 취소 API 구현

* feat(WaitingController): 예약 대기 조회 API 구현

* feat(AuthController): 로그아웃 API 구현

* refactor(/dto/response): 메서드 인자 줄바꿈 통일

* refactor(GlobalExceptionHandler): 출력문 대신 logger로 변경

* refactor(/repository): repository로 getBy 메서드 위치 이동

* chore(data.sql): 예약 대기 더미 데이터 생성

* docs(README.md): 4단계 요구사항 작성

* feat(ReservationController): 에약 대기 승인 기능 구현

* style(/js): 프론트 코드 추가

* feat(AdminController): 어드민 예약 대기 목록 조회 구현

* feat(AdminController): 예약 대기 거절 구현

* refactor(/util): AbstractTestExecutionListener 구현, Test 어노테이션에 반영

* refactor(AdminController): 관리자 예약 조회용 응답 dto 생성

* refactor(/dto): 안 쓰는 dto 제거

* test(AuthIntegrationTest): 로그아웃 테스트 추가

* fix(JwtTokenProvider): 회원의 이름을 위한 클래임 이름 변경

* refactor(MemberController): 내 예약 조회 메서드 이름을 변

- MemberIntegrationTest 테스트 개선

* style(/src): 불필요한 주석 제거

* style(/js): 관리자의 예약 조회 API 프론트 코드에서 변경

* fix(ReservationService): 예약 삭제 권한 로직 변경

- 테스트 개선

* test(ReservationServiceTest): 예약 관련 서비스 계층 테스트 추가

* refactor(CreateReservationRequest): 관리자에 의한 예약 추가 요청 dto 이름 변경

* test(WaitingIntegrationTest): 예약 대기 거절 실패 테스트 추가

* fix(WaitingService): 예약 삭제 시 업데이트를 위한 대기 삭제 메서드 추가

* docs(APISpecification.md): API 명세 최신화

* fix(waiting.js): admin API 변경에 따른 프론트 코드 변경

* refactor(ReservationTest): 테스트 메서드의 이름을 각 테스트 목적에 맞게 변경

* refactor(ReservationService): 예외 메시지를 자세하게 변경

* refactor(WaitingService): 동일한 역할의 메서드를 통일

* refactor(ReservationService): 메서드 명을 조금 더 추상화하여 수정

* refactor(MemberRepository): 예외 메세지를 내부에서 던지도록 변경

* refactor(BaseEntity): 엔티티 컬럼에 updatable, nullable 조건 추가

* refactor(/test/reservation/service): main과 동일하게 테스트 패키지 이동

* refactor(AdminWaitingController): 각 도메인의 책임에 맞게 위치를 변경 및 AdminWaitingController 생성

* refactor(/src/test): 각 테스트 명을 통일성있게 전체 변경

* style(/src): import 문 최적화

+++ src/main/java/roomescape/reservation/service/ReservationService.java
@@ -3,13 +3,16 @@ package roomescape.reservation.service;
 import java.time.LocalDate;
 import java.time.LocalTime;
 import java.util.List;
-import java.util.NoSuchElementException;
 import org.springframework.stereotype.Service;
 import roomescape.auth.domain.AuthInfo;
+import roomescape.common.exception.ForbiddenException;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
+import roomescape.reservation.dto.request.CreateMyReservationRequest;
+import roomescape.reservation.dto.request.CreateReservationByAdminRequest;
 import roomescape.reservation.dto.request.CreateReservationRequest;
 import roomescape.reservation.dto.response.CreateReservationResponse;
+import roomescape.reservation.dto.response.FindAdminReservationResponse;
 import roomescape.reservation.dto.response.FindAvailableTimesResponse;
 import roomescape.reservation.dto.response.FindReservationResponse;
 import roomescape.reservation.model.Reservation;
@@ -18,35 +21,56 @@ import roomescape.reservationtime.model.ReservationTime;
 import roomescape.reservationtime.repository.ReservationTimeRepository;
 import roomescape.theme.model.Theme;
 import roomescape.theme.repository.ThemeRepository;
+import roomescape.waiting.model.Waiting;
+import roomescape.waiting.repository.WaitingRepository;
+import roomescape.waiting.service.WaitingService;
 
 @Service
 public class ReservationService {
 
+    private final WaitingService waitingService;
     private final ReservationRepository reservationRepository;
     private final ReservationTimeRepository reservationTimeRepository;
     private final ThemeRepository themeRepository;
     private final MemberRepository memberRepository;
+    private final WaitingRepository waitingRepository;
 
-    public ReservationService(final ReservationRepository reservationRepository,
+    public ReservationService(final WaitingService waitingService,
+                              final ReservationRepository reservationRepository,
                               final ReservationTimeRepository reservationTimeRepository,
                               final ThemeRepository themeRepository,
-                              final MemberRepository memberRepository) {
+                              final MemberRepository memberRepository,
+                              final WaitingRepository waitingRepository) {
+        this.waitingService = waitingService;
         this.reservationRepository = reservationRepository;
         this.reservationTimeRepository = reservationTimeRepository;
         this.themeRepository = themeRepository;
         this.memberRepository = memberRepository;
+        this.waitingRepository = waitingRepository;
     }
 
-    public CreateReservationResponse createReservation(final AuthInfo authInfo,
-                                                       final CreateReservationRequest createReservationRequest) {
-        ReservationTime reservationTime = findReservationTime(createReservationRequest.timeId());
-        Theme theme = findTheme(createReservationRequest.themeId());
-        Member member = findMember(authInfo.getMemberId());
+    public CreateReservationResponse createMyReservation(final AuthInfo authInfo,
+                                                         final CreateMyReservationRequest createMyReservationRequest) {
+        CreateReservationRequest createReservationRequest = CreateReservationRequest.of(authInfo.getMemberId(),
+                createMyReservationRequest);
+        return CreateReservationResponse.from(createReservation(createReservationRequest));
+    }
+
+    public CreateReservationResponse createReservationByAdmin(
+            final CreateReservationByAdminRequest createReservationByAdminRequest) {
+        CreateReservationRequest createReservationRequest = CreateReservationRequest.of(createReservationByAdminRequest);
+        return CreateReservationResponse.from(createReservation(createReservationRequest));
+    }
+
+    public Reservation createReservation(final CreateReservationRequest createReservationRequest) {
+        ReservationTime reservationTime = reservationTimeRepository.getById(createReservationRequest.timeId());
+        Theme theme = themeRepository.getById(createReservationRequest.themeId());
+        Member member = memberRepository.getById(createReservationRequest.memberId());
 
         checkAlreadyExistReservation(createReservationRequest, createReservationRequest.date(), theme.getName(),
                 reservationTime.getStartAt());
         Reservation reservation = createReservationRequest.toReservation(member, reservationTime, theme);
-        return CreateReservationResponse.from(reservationRepository.save(reservation));
+        return reservationRepository.save(reservation);
     }
 
     private void checkAlreadyExistReservation(final CreateReservationRequest createReservationRequest,
@@ -60,35 +84,21 @@ public class ReservationService {
         }
     }
 
-    private ReservationTime findReservationTime(final Long id) {
-        return reservationTimeRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("식별자 " + id + "에 해당하는 시간이 존재하지 않아 예약을 생성할 수 없습니다."));
-    }
-
-    private Theme findTheme(final Long id) {
-        return themeRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("식별자 " + id + "에 해당하는 테마가 존재하지 않아 예약을 생성할 수 없습니다."));
-    }
-
-    private Member findMember(final Long id) {
-        return memberRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("식별자 " + id + "에 해당하는 회원이 존재하지 않아 예약을 생성할 수 없습니다."));
-    }
-
-    public List<FindReservationResponse> getReservations() {
+    public List<FindAdminReservationResponse> getReservations() {
         return reservationRepository.findAll().stream()
-                .map(FindReservationResponse::from)
+                .map(FindAdminReservationResponse::from)
                 .toList();
     }
 
     public FindReservationResponse getReservation(final Long id) {
-        Reservation reservation = findReservation(id);
+        Reservation reservation = reservationRepository.getById(id);
         return FindReservationResponse.from(reservation);
     }
 
-    private Reservation findReservation(final Long id) {
-        return reservationRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("식별자 " + id + "에 해당하는 예약이 존재하지 않아 예약을 조회할 수 없습니다."));
+    public List<FindReservationResponse> getReservations(final AuthInfo authInfo) {
+        return reservationRepository.findAllByMemberId(authInfo.getMemberId()).stream()
+                .map(FindReservationResponse::from)
+                .toList();
     }
 
     public List<FindAvailableTimesResponse> getAvailableTimes(final LocalDate date, final Long themeId) {
@@ -114,21 +124,28 @@ public class ReservationService {
                 .toList();
     }
 
-    public void deleteReservation(final Long id) {
-        validateExistReservation(id);
+    public void deleteReservation(final AuthInfo authInfo, final Long id) {
+        Reservation reservation = reservationRepository.getById(id);
+        checkCancelAuthorization(reservation, authInfo.getMemberId());
+
+        if (waitingRepository.existsByReservation(reservation)) {
+            updateReservationByMember(reservation);
+            return;
+        }
         reservationRepository.deleteById(id);
     }
 
-    private void validateExistReservation(final Long id) {
-        if (!reservationRepository.existsById(id)) {
-            throw new NoSuchElementException("식별자 " + id + "에 해당하는 예약이 존재하지 않습니다. 삭제가 불가능합니다.");
+    private void checkCancelAuthorization(final Reservation reservation, final Long memberId) {
+        Member member = memberRepository.getById(memberId);
+        if (member.isNotAdmin() && !reservation.isOwnedBy(member)) {
+            throw new ForbiddenException(
+                    "식별자 " + reservation.getId() + "인 예약에 대해 회원 식별자 " + memberId + "의 권한이 존재하지 않아, 삭제가 불가능합니다.");
         }
     }
 
-    public List<roomescape.member.dto.response.FindReservationResponse> getReservationsByMember(
-            final AuthInfo authInfo) {
-        return reservationRepository.findAllByMemberId(authInfo.getMemberId()).stream()
-                .map(roomescape.member.dto.response.FindReservationResponse::from)
-                .toList();
+    private void updateReservationByMember(final Reservation reservation) {
+        Waiting waiting = waitingRepository.getFirstByReservation(reservation);
+        reservation.updateMember(waiting.getMember());
+        waitingService.deleteWaitingForReservationUpgrade(waiting.getId());
     }
 }
